Updated build configs
removed rti.c and rti.h - they were not being used and were disabled in HCG. This solves the issue of having to add the #endif for vPortPreemptiveTick that was present in rti.c. It is only required for the launchpad port. 

Also added flash new transfer group, increased flash speed to 2 MHz for v0.3. Unchanged at 400 KHz for launchpad, and is super fast for 0.4.