SPIFFs reliability mostly fixed
The issue was that we'd get into a weird state where we couldn't read because writes were taking a while since the FS was moving stuff around. Solution: add another top level mutex such that we cannot interrupt a write routine with a read (and vice versa). 

It doesn't appear that the built in mutex handles are going to work. This two-level mutex fixes that.