Figured out spiffs stability issue - stack
The issue was printf using too much stack. Max task stack is allocated at task creation time, and comes from the RTOS heap. 

RTOS heap comes from FreeRTOSConfig.h. We have about 4k of overhead, and the rest of the 127k of RAM can probably be allocated to the RTOS.