Fixed bug where first files wouldn't write
Before, we would successfully create the first files and write "created" to them, but any subsequent writes to the aX files would fail, but bX writes would be fine, as would the next aX writes after that. 

To fix, sfu_spiffs_init() was moved into the fs lifecycle task, and called right before file creation. Before, it was run at the top of task_main. 

Not exactly sure /why/ it works now, but it seems that running things before the while loop in a task can cause issues. It's easier to spawn a task that just runs the function, or do something like we're doing now, and run in the body of a task.